Temporarily revert default dll characteristics.
Until more testing is done, this patch reverts the
DEFAULT_DLL_CHARACTERISTICS to 0, but keeps the new --disable- options
for the flags so that they can be overridden later on the command line.

Note that this does NOT revert the default --enable-reloc-section.  This
shouldn't hurt anything other than making the resulting executable
larger.